+++ scripts/build.mjs
@@ -28,3 +28,4 @@ await copyFile(new URL('src/image-import.js', root), new URL('image-import.js', 
 await copyFile(new URL('src/table-import.js', root), new URL('table-import.js', dist));
 await copyFile(new URL('src/table-cell-import.js', root), new URL('table-cell-import.js', dist));
 await copyFile(new URL('src/table-border-import.js', root), new URL('table-border-import.js', dist));
+await copyFile(new URL('src/color-ref.js', root), new URL('color-ref.js', dist));